Adds build URL to end of Slack message if defined
Define the `build_env` key inside of vidar.yml to enable this feature.

  ---
  build_env: "TRAVIS_BUILD_WEB_URL"

When this configuration key is set, the value is used to get the ENV
variable and use it in the Slack notification.  If it is not set, or is
set to an empty string, then the Slack notification does not include the
link.